Move the new functions to snappi_tests/files/helper.py instead of the common folder.
