+++ tests/common/snappi_tests/snappi_fixtures.py
@@ -19,12 +19,8 @@ from tests.common.helpers.assertions import pytest_assert
 from tests.snappi_tests.variables import dut_ip_start, snappi_ip_start, prefix_length, \
     dut_ipv6_start, snappi_ipv6_start, v6_prefix_length, pfcQueueGroupSize, \
     pfcQueueValueDict          # noqa: F401
-from tests.snappi_tests.variables import MULTIDUT_PORT_INFO, MULTIDUT_TESTBED
-from tests.snappi_tests.files.helper import skip_warm_reboot
-from tests.common.config_reload import config_reload
-from tests.common.reboot import reboot
-from tests.common.helpers.parallel import parallel_run
-from tests.common.utilities import wait_until
+
+
 logger = logging.getLogger(__name__)
 
 
@@ -1339,80 +1335,3 @@ def check_fabric_counters(duthost):
                               format(crc_errors, duthost.hostname, val_list[0], val_list[1]))
                 pytest_assert(fec_uncor_err == 0, 'Forward Uncorrectable errors:{} for DUT:{}, ASIC:{}, Port:{}'.
                               format(fec_uncor_err, duthost.hostname, val_list[0], val_list[1]))
-
-
-@pytest.fixture(autouse=True, params=MULTIDUT_PORT_INFO[MULTIDUT_TESTBED])
-def multidut_port_info(request):
-    yield request.param
-
-
-@pytest.fixture(autouse=True)
-def setup_ports_and_dut(
-        duthosts,
-        snappi_api,
-        get_snappi_ports,
-        multidut_port_info,
-        number_of_tx_rx_ports):
-    for testbed_subtype, rdma_ports in multidut_port_info.items():
-        tx_port_count, rx_port_count = number_of_tx_rx_ports
-        if len(get_snappi_ports) < tx_port_count + rx_port_count:
-            pytest.skip(
-                "Need Minimum of 2 ports defined in ansible/files/*links.csv"
-                " file, got:{}".format(len(get_snappi_ports)))
-
-        if len(rdma_ports['tx_ports']) < tx_port_count:
-            pytest.skip(
-                "MULTIDUT_PORT_INFO doesn't have the required Tx ports defined for "
-                "testbed {}, subtype {} in variables.py".format(
-                    MULTIDUT_TESTBED, testbed_subtype))
-
-        if len(rdma_ports['rx_ports']) < rx_port_count:
-            pytest.skip(
-                "MULTIDUT_PORT_INFO doesn't have the required Rx ports defined for "
-                "testbed {}, subtype {} in variables.py".format(
-                    MULTIDUT_TESTBED, testbed_subtype))
-        logger.info('Running test for testbed subtype: {}'.format(testbed_subtype))
-        snappi_ports = get_snappi_ports_for_rdma(
-            get_snappi_ports,
-            rdma_ports,
-            tx_port_count,
-            rx_port_count,
-            MULTIDUT_TESTBED)
-        testbed_config, port_config_list, snappi_ports = snappi_dut_base_config(
-            duthosts, snappi_ports, snappi_api, setup=True)
-
-    if len(port_config_list) < 2:
-        pytest.skip("This test requires at least 2 ports")
-    yield (testbed_config, port_config_list, snappi_ports)
-
-    snappi_dut_base_config(duthosts, snappi_ports, snappi_api, setup=False)
-
-
-@pytest.fixture(params=['warm', 'cold', 'fast'])
-def reboot_duts(setup_ports_and_dut, localhost, request):
-    reboot_type = request.param
-    _, _, snappi_ports = setup_ports_and_dut
-    skip_warm_reboot(snappi_ports[0]['duthost'], reboot_type)
-    skip_warm_reboot(snappi_ports[1]['duthost'], reboot_type)
-
-    def save_config_and_reboot(node, results=None):
-        logger.info("Issuing a {} reboot on the dut {}".format(reboot_type, node.hostname))
-        node.shell("mkdir /etc/sonic/orig_configs; mv /etc/sonic/config_db* /etc/sonic/orig_configs/")
-        node.shell("sudo config save -y")
-        reboot(node, localhost, reboot_type=reboot_type, safe_reboot=True)
-        logger.info("Wait until the system is stable")
-        wait_until(180, 20, 0, node.critical_services_fully_started)
-
-    # Convert the list of duthosts into a list of tuples as required for parallel func.
-    args = set((snappi_ports[0]['duthost'], snappi_ports[1]['duthost']))
-    parallel_run(save_config_and_reboot, {}, {}, list(args), timeout=900)
-
-    yield
-
-    def revert_config_and_reload(node, results=None):
-        node.shell("mv /etc/sonic/orig_configs/* /etc/sonic/ ; rmdir /etc/sonic/orig_configs; ")
-        config_reload(node, safe_reload=True)
-
-    # parallel_run(revert_config_and_reload, {}, {}, list(args), timeout=900)
-    for duthost in args:
-        revert_config_and_reload(node=duthost)
